Unify storage tests in storage test project
