+++ src/Immediate.Jobs.EntityFrameworkCore/Immediate.Jobs.EntityFrameworkCore.csproj
@@ -32,7 +32,6 @@
   </ItemGroup>
 
   <ItemGroup>
-	<InternalsVisibleTo Include="Immediate.Jobs.FunctionalTests" />
 	<InternalsVisibleTo Include="Immediate.Jobs.StorageTests" />
   </ItemGroup>
