+++ src/Immediate.Jobs.Shared/Immediate.Jobs.Shared.csproj
@@ -20,6 +20,7 @@
   <ItemGroup>
 	<InternalsVisibleTo Include="Immediate.Jobs.Benchmarks" />
 	<InternalsVisibleTo Include="Immediate.Jobs.FunctionalTests" />
+	<InternalsVisibleTo Include="Immediate.Jobs.StorageTests" />
   </ItemGroup>
 
 </Project>
