+++ tests/Immediate.Jobs.FunctionalTests/Immediate.Jobs.FunctionalTests.csproj
@@ -15,8 +15,6 @@
   <ItemGroup>
 	<PackageReference Include="Immediate.Handlers" />
 	<PackageReference Include="Microsoft.AspNetCore.TestHost" />
-	<PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" />
-	<PackageReference Include="SQLitePCLRaw.lib.e_sqlite3" />
     <PackageReference Include="Microsoft.Testing.Extensions.CodeCoverage" />
 	<PackageReference Include="Verify.XunitV3" />
     <PackageReference Include="xunit.v3.mtp-v2" />
@@ -27,7 +25,6 @@
 	                  OutputItemType="Analyzer"
 	                  ReferenceOutputAssembly="false" />
 	<ProjectReference Include="../../src/Immediate.Jobs.Dashboard/Immediate.Jobs.Dashboard.csproj" />
-	<ProjectReference Include="../../src/Immediate.Jobs.EntityFrameworkCore/Immediate.Jobs.EntityFrameworkCore.csproj" />
 	<ProjectReference Include="../../src/Immediate.Jobs.NodaTime/Immediate.Jobs.NodaTime.csproj" />
     <ProjectReference Include="../../src/Immediate.Jobs.Testing/Immediate.Jobs.Testing.csproj" />
     <ProjectReference Include="../../src/Immediate.Jobs.Generators/Immediate.Jobs.Generators.csproj"
