+++ tests/Immediate.Jobs.StorageTests/JobExecutionStorageTests.cs
@@ -2,7 +2,7 @@ using Immediate.Jobs.Shared.Apis;
 using Immediate.Jobs.Shared.Storage;
 using Microsoft.Extensions.Time.Testing;
 
-namespace Immediate.Jobs.FunctionalTests.Storage;
+namespace Immediate.Jobs.StorageTests;
 
 public sealed class JobExecutionStorageTests
 {
